requirements.txt
streamlit
requests
python-dotenv

python-docx>=0.8.11
PyPDF2>=3.0.0
googletrans>=3.1.0a0
langdetect>=1.0.9
pandas>=1.3.0
tqdm>=4.62.0
beautifulsoup4>=4.9.3
fake-useragent>=0.1.11
datasets>=2.18.0 
transformers>=4.37.0
torch>=2.2.0
accelerate>=0.27.0 